[FIX] whatsapp: prefer report over static header document
When the user sends a template in which the header type is set as a document
containing both the report and the static header, the static header will be sent
rather than the report.

Steps to produce:-

- Create a template with header type document.
- Add a static header attachment and a report too.
- Send a whatsapp message with created template.
- Static header will be sent rather than report.

Problem:

A default header attachment is added to templates for header-type images,
videos, and documents as it is a required field while syncing the templates.
Now, the user adds a report to that template and tries to send it then static
header is sent over report.

Solution:

Whenever the user adds a report to a template then static header will be
invisible and the report will be sent in the message. If the user changes the
model of the template in which the report is set then the report will be droped.

Task-3640336 (WhatsApp: Fix Document Header / Report mismatch)

X-original-commit: 987b830240227be51605882dd67b439b5ba7080c
Part-of: odoo/enterprise#60098