[FIX] l10n_cl_edi: Adjust discount handling in fetchmail XML parsing
fix line discounts giving priority for discount amount instead of discount percentage
add surcharge amount in the lines.
Adjust discount handling in fetchmail XML parsing

There have been detected several discrepancies between the discount amounts
in the XML files read by fetchmail and the calculated discount in Odoo.
In Chile, line item amounts do not use decimals, and Odoo represents
discounts as a percentage. This leads to a mismatch when calculating the discount
percentage by dividing the discount by the amount and multiplying by 100.

As a result, the total amount calculated by Odoo, the net amount, or the tax
amount often differs from the official record in the SII.
To simplify the logic and reduce the possibility of errors, we have decided
to deduct the discount from the unit price instead of calculating and
applying it as a discount on the invoice lines."

[FIX] l10n_cl_edi: add feature to deduct the VAT when tax is included in XML
values (MntBruto set to 1).
Add a test for an invoice with discounts and amounts with tax included

closes odoo/enterprise#67608

X-original-commit: b01eabcf08bf24e33d8624cf0c1ee9b8371fd888
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>