[FIX] l10n_fr_fec_import: don't create mail aliases by default
It is possible that a FEC file contains journals with different codes but same name. In such case, each of these would try to create a distinct mail alias with the same name ; this is not allowed and raised an error. We now prevent that by not generating any mail alias by default on journals imported via FEC. They can still be added later on by the user if he wishes to do so.

OPW 3813584

closes odoo/enterprise#59921

X-original-commit: 430cdcc06d29296cd6b4db992db93c1e2b527cdb
Related: odoo/odoo#160168
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>