[FIX] l10n_fr_fec_import: don't create mail aliases by default
It is possible that a FEC file contains journals with different codes but same name. In such case, each of these would try to create a distinct mail alias with the same name ; this is not allowed and raised an error. We now prevent that by not generating any mail alias by default on journals imported via FEC. They can still be added later on by the user if he wishes to do so.

OPW 3813584

closes odoo/enterprise#60020

X-original-commit: 8297f4454309853cb6d6afad853209f06fb0b60f
Related: odoo/odoo#160374
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>