[IMP] knowledge: remove hack to update the title of the embedded view
Due to time constraint, we didn't have time to cleanly implement the
"rename" feature of the embedded views when we released the feature.
We directly patched the DOM by using a query selector. Of course, this
is not a good practice as the view component can be re-rendered and the
modification can be lost.

To fix that issue, we will now call the `setDisplayName` method defined
in the environment of the embedded view. We will then no longer use a
query selector to manually patch the DOM.

task-3381729

Part-of: odoo/enterprise#51284