[FIX] delivery_sendcloud: use correct shipping price
Before this commit, the lowest shipping rate was used as the
delivery cost, as in: https://github.com/odoo/enterprise/blob/8c40ba77b23c4ccd24c3785bc1e6cf5b3fbe0e51/delivery_sendcloud/models/sendcloud_service.py#L91
However, the prices received from the Sendcloud are strings,
and as a result, this line returned the rate with the lowest
lexicographical order.

As a result, sometimes the rates with a higher value were
chosen. The issue happens when a shipping product costs a
single digit amount (e.g. `9.00`) but a second product costs
in double digits (e.g. `11.00`). In this case the latter is
used which is wrong.

This commit, converts the prices to float before comparing
them to get the actual best price.

opw-3815117

closes odoo/enterprise#59536

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>