[REF] web_enterprise, *: simplify the color_scheme service
*: website_enterprise

This commit removes unnecessary complexity to the color_scheme service.
Most of those changes were introduced in commit (1) to improve the experience
when navigating to Studio, since dark mode was not supported.

After commit (2), Studio module supports dark mode fully, and the patch of the
service was no longer necessary. This commit removes the need for the extra
cookie 'configured_color_scheme', and simplifies the service once again.

1) https://github.com/odoo/enterprise/commit/fda6f9cc47b9f9849155032fc0c53983a561e15d
2) https://github.com/odoo/enterprise/commit/f71d2a04fdada581c2e23776883c6fda6a21b8da

closes odoo/enterprise#59597

Signed-off-by: Romain Estievenart (res) <res@odoo.com>