[FIX] hr_contract_salary: set correct job_id
Since when we generate an offer, we can provide a specific job and
department, that will be set on the user and used in the pdf contract
signed.

We should also get this job and departement correctly set on the new
contract which is not the case for now.

So we first choose to use the info comming from the offer before the one
from the contract template.

closes odoo/enterprise#59945

X-original-commit: b52e5a34469380ad39745cc07b4b1385e39c7042
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>