[FIX] hr_contract_salary: set correct job_id
Since when we generate an offer, we can provide a specific job and
department, that will be set on the user and used in the pdf contract
signed.

We should also get this job and departement correctly set on the new
contract which is not the case for now.

So we first choose to use the info comming from the offer before the one
from the contract template.

closes odoo/enterprise#59882

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>