[IMP] sign: merge refused and ignored to cancel state
Although we make it possible to opt-out of automated reminders via a link in
the email, this is not super visible and we should also make this possible in
the sign UI. To simplify some flows, we can merge this with the 'refusal' flow
and with the 'ignore' flow. Indeed, if one decides to ignore a sign request it
means they have no intent to sign - might as well cancel it.

Essentially, this commit merges refused and ignored states to canceled in the
backend. On top of that we also add a cancel button on the kanban view of sign
request and on the control panel in the sign request client action.

task-3675303

Part-of: odoo/enterprise#55463