[IMP] documents_spreadsheet: redirect share to Documents app
Currently when sharing a spreadsheet by using the share option or copying the
URL, as an internal user, you don't land in the Documents app.

With this commit, the user lands on the open spreadsheet, with the
Documents app in the breadcrumb, as if he was coming from Documents.

closes odoo/enterprise#59564

Task: 3525858
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>