[FIX] account_reports: Don't use the load more limit when exporting file
The aim of this commit is avoiding using the load more limit field
from the report when we are exporting the report as file.
e.g.: Users want to download their Partner VAT listing in Belgium,
before this commit, the report will follow the load more limit
sets on the report, after this fix, not.

no task id

Part-of: odoo/enterprise#59573