[FIX] pos_restaurant_appointment: appointment end of day
Before this commit, if the end of a table appointment was ending
the day after the start day, the appointment wouldn't
be displayed. Fixing this, a new issue arises, if an
appointment start the day before, showing it next day
will create confusion since it will display the starting
time of the previous day. To fix this we simply set the
starting time to the start of the day.

closes odoo/enterprise#59761

X-original-commit: 09f57d2bc0832368ee3be11bf000ae5f23ac7c19
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>