[IMP] sign: Improve the onboarding experience
Before this commit, the Sign app's onboarding process needs to be smoother for
better customer experience and ease of use.

After this commit, users are allowed to upload templates with ease by simply
drag and dropping files. The action helper has been updated to give a more
Intuitive and seamless experience. Fine-tuning menu sequences and removing
seconds from the datetime field in the Kanban view resulted in an improvement in
overall usability.

task-3764704

closes odoo/enterprise#60813

Related: odoo/upgrade#6003
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>