[FIX] delivery_ups_rest: crash when alert has no description
Sometimes UPS return a list of alerts. In that case the code works
since it's the type we expect.

But when there is only one alert UPS doesn't return a list but a
single object. In that case we ierate on the keys and it crashs because
the first key is code and not description

closes odoo/enterprise#59999

X-original-commit: 41c89e7b0d5dacda3caf37fd8e6e17dcad3fb0ad
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>