[FIX] knowledge: block the readonly user from the history
This commit fixes an issue where readonly users are able to access and restore
an old version of an article via the history dialog but it didn't affect the DB
as no write could be done by this user.
The change was only visible and functionnally didn't change the body field.

Now when a user accesses to an article where he can only read, the history
dialog button is hidden.

task-3836201

closes odoo/enterprise#60500

X-original-commit: e87ea7cd5788274af57ddfdc50ce3c0901f9636b
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>