[IMP] planning: Undo shift splitting in gantt view of Planning
After this commit, a notification appears when the user uses the split
tool to divide a shift in the gantt view. The notification contains an "Undo"
button which allows to merge back the last divided shift.
When "Undo" is pressed, a confirmation notification is displayed to tell
that the shift was successfully merged back.

task-3827618

closes odoo/enterprise#59620

Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>