[FIX] appointment_hr: update query count due to menu cache change
[This first commit] fixed an issue with the website menu cache. The menu
is not cached anymore if there is a record like URL in the it. When
`website_helpdesk` module is installed, since `saas-16.3`, the menu
contains a record like URL which disable the menu cache and increase the
number of queries. The solution is to increase the max number of queries
even if a better solution would be to remove the `Help` link from the
menu but we can't do that because `appointment_hr` doesn't depend on
`website`.

[This first commit]: https://github.com/odoo/odoo/commit/43576cd424b6d0fc7da01142b5e6550e371ad1ff

runbot-60956
runbot-59981

closes odoo/enterprise#59954

X-original-commit: 7d1d0b2e703294e772678822c54926b5e6ee8957
Signed-off-by: Romain Derie (rde) <rde@odoo.com>
Signed-off-by: Guillaume Dieleman (gdi) <gdi@odoo.com>