[FIX] knowledge: make embedded views scrollable
This commit fixes an issue with the embedded views of the knowledge article
view: the embedded could take as much height as they needed which could
result in excessive place taken by a single embedded view and this also
had issues with the virtual hook which is not adapted to embedded views
without scroll (this could result in completly blank space inside
embedded gantt views when they would take too much space). To solve
this issue, the commit adds a fixed maximum height to the embedded views
(80% of the window height).

task-3834795

closes odoo/enterprise#60117

X-original-commit: 0ec4dc6bc8caa51c923dcabbf41d8fe02c471387
Related: odoo/odoo#160568
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>