[FIX] l10n_be_reports: remove group_by on BE_PL_14P report line
In the "Profit and Loss" report, the "14P - Profit (Loss) of the Preceding
Period Brought Forward (+)/(-)" line is grouped by account_id, showing then
all the accounts from the preceding period that impact the final amount.
Which provides a lot of information that may be too much.
Only the total amount should be displayed.

opw-3810161

closes odoo/enterprise#59640

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>