[REV] project_timesheet_forecast: don't leave out leaves
This reverts commit 55cba9f899be0f058834bcd81bd239547c9cebd6.

The issue this intended to fix was rooted in the fact that no timesheets
were created for public holidays, leading to the remaining hours shown
in Timesheets & Planning analysis being wrong.

This is no longer an issue starting from version 16.3, timesheets do get
created, so any *planned* hours that fall on a holiday gets subtracted
by the corresponding effective hours from timesheets.

opw-3509155

closes odoo/enterprise#59707

X-original-commit: eeb80ac5ece0bbd5c221ce1883ec58609ca28184
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>