[FIX] industry_fsm_sale: add correct context key for invoice button
Steps to reproduce issue:

1. Go to _Field Service > All Tasks > To Invoice_
    - If no task is shown, remove To Invoice filter
2. Click on the checkbox of a task
3. The button Create Invoice should appear next to the Print and Actions buttons but it does not

Explanation:

The context key used to show the element contains a `search_default` prefix. Keys with this prefix are specifically used to create a search filter and are deleted in the process.
https://github.com/odoo/odoo/blob/71c81e605b2f06889ad8031f021863638edab268/addons/web/static/src/search/search_model.js#L291-L300

Suggested fix:

No existing key in the context is specific to the action nor is their value. Adding a new specific context key prevents unintended behaviours.

opw-3773005

closes odoo/enterprise#60044

X-original-commit: 612d22a57ce8f1f8489660e45a0cd376bebb5c7c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Stroobant Paul (stpa) <stpa@odoo.com>