[FIX] whatsapp: copy attachments when copying templates
If you duplicate a template using document header and an attachment, this
latter one is not copied which leads to an invalid template.

As attachments is a m2m relationship, we cannot just use copy=True as it would
copy a reference to the attachment. Instead of duplicated the attachment itself
so that original and clone templates have different ir.attachment records.

We also fix a missing trigger in the check about attachments or report_id as
it was possible to create invalid templates, by removing report_id. Check
on attachment or report existence was not done.

Task-3605612 (WhatsApp: Fix header / upload / reporting usage)

X-original-commit: odoo/enterprise@56ceb2ee72196aec5d949015769436f77cfb4540
Part-of: odoo/enterprise#59987