[FIX] whatsapp: limit document header to 1 attachment or report
To be sure either a single attachment either a report is defined on a template
we use the newly-introduced option of m2m binary widget to limit number of
available file upload to 1.

Report field is now also hidden when there are attachments, like attachments
are hidden when a report is set.

Task-3605612 (WhatsApp: Fix header / upload / reporting usage)

X-original-commit: 3f524dcd68544ef3a0f493321b24125b2f30e895
Part-of: odoo/enterprise#60098