[FIX] whatsapp: limit document header to 1 attachment or report
To be sure either a single attachment either a report is defined on a template
we use the newly-introduced option of m2m binary widget to limit number of
available file upload to 1.

Report field is now also hidden when there are attachments, like attachments
are hidden when a report is set.

Task-3605612 (WhatsApp: Fix header / upload / reporting usage)

X-original-commit: odoo/enterprise@94b355681d8b48629a3ab18f781c19fdd2dd15a1
Part-of: odoo/enterprise#59987