[FIX] l10n_br_sale_subscription: hide non-tax amounts again
4755b82fa40df50362e72ccc3e08f6ce5383dde3 changed the portal template
used for subscriptions by overriding
`_get_name_portal_content_view()`. Because of this, the override in
`l10n_br_sales` no longer applies and we end up with the standard
subscription portal view instead of the customized one for Brazil.

Although not very nice, the only way around it I could think of was to
create this new module so we can override again with a new portal
template that inherits from
sale_subscription.subscription_portal_content.

closes odoo/enterprise#60388

X-original-commit: af3f12436f317d29c93f38eb4a2b00fbbf1af3a8
Related: odoo/odoo#161232
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>