[FIX] account_online_synchronization: skip online bank sync update when opening journal
When opening journal, a check is done to update the bank sync state
It is not possible to update the state without accounting manager access rights

Introducing a skip for non account managers to allow opening the journal. Furthermore, non-managers can't see the online sync anyway so we don't update it until a manager opens the journal.

opw-3800147

closes odoo/enterprise#59706

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>