[FIX] sale_ebay: restrict use of attributes ebaysdk connection
The API of the ebaysdk tends to change from one version to
another to another, and developers tend to not install the
library version pinned in requirements.txt.

To avoid compatibility patches for one version
to another, limit on purpose attributes and methods
which can be use to the bare minimum of what we currently use,
to discourage using new methods.

Part-of: odoo/enterprise#47558