[IMP] stock_barcode: added a sound when a barcode scan is successful
Before this commit
===============
When a barcode was scanned successfully, the screen flashed to indicate the
successful scan, but there was no sound. This lack of auditory feedback could
sometimes make it difficult for users to be sure the scan was successful.

After this commit
===============
Along with the flashing screen, a sound now plays when a barcode scan is
successful. This provides clear and immediate feedback to users, ensuring they
are aware that the scan was successful.

closes odoo/enterprise#59908

Taskid: 3685476
Signed-off-by: William Henrotin (whe) <whe@odoo.com>