[FIX] knowledge: use search model of imported view
**Step to reproduce:**

- Install `Knowledge` and `Documents` modules (for test purpose)
- Go to Documents app
- Click on Action button -> Knowledge -> Insert view in article
- Select any article

**Issue:**

View is not loaded correctly in the article.
Error message: Something went wrong! The view does not exist or you
               are not allowed to access to it.

**Cause:**

In the imported view, we try to retrieve the selected Folders, but this
last one need the search model to retrieve the data.

https://github.com/odoo/enterprise/blob/11daa694d4ca51939c2b458bced8ab9d48030ef8/documents/static/src/views/documents_controller_mixin.xml#L5

The method called to retrieve the folders (`getSelectedFolder`) is set
on the documents search model (`DocumentsSearchModel`), however, the
search model set in the embedded_view is the knowledge search model
(`KnowledgeSearchModel`).

**Solution:**

Change the `KnowledgeSearchModel` class into a mixin so it can extend
the (potentially) custom search model defined for any embedded view
during the mounting process (default to the generic search model),
instead of overwriting it.

opw-3752927

closes odoo/enterprise#59841

X-original-commit: 396f1e752c66d9cfb8a7521ed576be9a25900f27
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Nasreddin Boulif <bon@odoo.com>