[FIX] documents: don't create a document twice on file upload
**Steps to reproduce:**

- Install `documents_project` module (for test purposes)
- Create a new activity type with:
  - `action`: `Upload Document`
  - `folder_id`: `Internal`
  - `model_id`: `Task`
- Go to settings and enable `Centralize files attached to projects
and tasks` (from 16.0, set a default folder on the project)
- Go to any task and add a new activity with the new activity type
- Click on `Upload Document` and select a file

**Issue:**

  Access right error message.

**Cause:**

Simple explanation:
2 documents are created (one on the activity creation and one on the
attachment upload) and the attachment is linked to the second document.
When trying to fisrt unlink the second document and then link the
attachment to the first document, the attachment is already unlinked.

Detailed explanation:
When creating the activity, if a folder_id is set on the activity, an
empty document is first created with the request_activity_id.
https://github.com/odoo/enterprise/blob/a23a18681bddc5995c5dac8cacfb074c06fc5ea8/documents/models/mail_activity.py#L46

When uploading the file, if the related record model is an inherit of
`documents.mixin` model and the documents settings (in this case
`Document Project Settings`) are activated, it will create the
document with the attachment.
https://github.com/odoo/enterprise/blob/2df654e8cb08d528976d5f1d24397574798a5cad/documents/models/ir_attachment.py#L63

Then, in the action done, we will unlink the last document record
created and try to link its attachment (already unlinked) to
the first document record (with the `request_activity_id`).
https://github.com/odoo/enterprise/blob/2df654e8cb08d528976d5f1d24397574798a5cad/documents/models/mail_activity.py#L33

Since the following commit, when unlinking a document, it unlink also
it's attachment: https://github.com/odoo/enterprise/commit/a999f2c32ab542ca7aa44cf34970dc7cca4fdaf8

**Solution:**

Override the upload route (`/mail/attachment/upload`) so that it skip
the creation of the second document (by adding `no_document` to the
context`) if an activity ID is available and that a document with a
`request_activity_id` with the same activity ID already exists.

COM PR: https://github.com/odoo/odoo/pull/159943

opw-3458850

closes odoo/enterprise#63312

X-original-commit: bdea145823b266b76ae276ae0f096590cdd1e437
Related: odoo/odoo#166964
Signed-off-by: Nasreddin Boulif (bon) <bon@odoo.com>