[FIX] pos_iot: reset is_posbox flag at uninstall
In `pos_enterprise` we alias `is_posbox` as `module_pos_iot`. If the
value is True we install the module `pos_iot` when a write in pos.config
model happens:

The issue is that when we uninstall `iot` module the flag is never
reset. This then causes issues if we try to install a module that writes
on any value of pos.config.

closes odoo/enterprise#59949

X-original-commit: 6b10cc80ea2441b5b2ab86aab52abbf7084d4319
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>