[FIX] l10n_fr_fec_import: raise UserError if key not found
When users try to import a file into the "FEC import" and the file doesn't
have the value of "key", an error occurs. This happens because the system
requires the value of "JournalCode" to be present in the imported file for
successful processing.

Steps to reproduce:
- Install l10n_fr_fec_import module
- Change company from YourCompany to FR Company
- Now Accounting -> Configuration -> Accounting -> Journals
- Select all journals -> export all journals -> download file in CSV format.
- Configuration -> Settings -> Accounting Import -> import ->
  Click on the "Import FEC button"
- Upload a file that you have downloaded and Import

Traceback: AttributeError
'NoneType' object has no attribute 'replace'

This commit resolves the mentioned issue by verifying if the key is present;
otherwise, it will raise a UserError. Apart from this, I have also added a .pot
and .po file for this module.

sentry - 4929578488

closes odoo/enterprise#59997

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>