[FIX] account_reports: reduce partner ledger memory usage
Issue -->

Method `_query_partners` does a search on `res.partner` which is used in the
return value. This return value is then looped over in the method `_build_partner_dicts`
to create the rows on the partner ledger. If the number of returned
`res.partner` records exceed a certain number, there is an overutilzation of
memory when getting fields id, name and trust in method `_get_report_line_partners`
leading to a MemoryError when loading the partner ledger report.

Solution -->

Replace the `search` call with a `search_fetch` call and store the specific
field values in the return value on `_query_partners`. This prevents the fetch
of the field values.

opw-3785731

closes odoo/enterprise#60148

X-original-commit: 8133b2bbcad42c87e857b18bba622f5c88fca75a
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>