[FIX] timesheet_grid: prevent infinite render loop
Steps to reproduce:
-------------------
- in My Timesheet, click on start;
- add a project, a task and a description;
- click on search icon in the grid (magnifying glass);

Issue:
------
The list view never opens.

Cause:
------
When the component performs the `onWillStart` hook, we add the
timesheet with the current timer and put it into edit mode.
This will have the effect of making a new render
(destroy the component and create a new one).
This will create a new component and repeat the process.

Solution:
---------
Let the component load the timesheet during the `onWillStart`
but manage the edit mode during the `onMounted` hook.

opw-3788661

closes odoo/enterprise#59896

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>