[FIX] knowledge: fix access denied error for thread tour
This commit fixes an issue with one of the nightly runbot test: No Demo.
This trigger launches tests without using any demo data for all of the modules
the issue with it is that the user `demo`, which is used in some tours, is part
of the demo data.

This leads to the thread tour to fail since it is using a demo user.
To fix this, the commit creates a new internal test user inside the test class
which is then used as login when launching the tour.

This way the tour is no longer dependant on a demo user without using the admin
login.

task-3844899

closes odoo/enterprise#60095

X-original-commit: 32b9357bc4b6b9e0da94ad5e79ca647ec7299bd0
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>