[FIX] knowledge: use more specific selector for main article emoji
To stylish the main emoji of the article, we used the CSS selector
`.o_field_knowledge_icon`. However, this selector is too generic and
will match with other emoji fields present in the embedded views
(see: `knowledge_icon` field).

Example: In 17.0, we re-used the emoji field in the embedded kanban
views of article items. With the selector `o_field_knowledge_icon`,
those emojis will be as big as the main emoji, which is not what we want.

To ensure that the CSS rules only applies to the main emoji of the
article, we will instead use the selector `o_knowledge_icon`, which is
specific to the main emoji of the article.

Related: https://github.com/odoo/enterprise/pull/54522

task-3784013

closes odoo/enterprise#60474

X-original-commit: 4578912f4de5c16143210bdbe944550a7db317cf
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>