[FIX] l10n_nl_reports_sbr{_icp}: switch template to newer nt18
The templates are used to create a report that will be sent to the
fiscal authorities of the Netherlands. They have updated to a newer
version of the template.
As it might happen again in the future, and the customer needs to
upgrade their module to get the updated templates, we add a warning
to ask them to upgrade before submitting a ticket.

closes odoo/enterprise#60132

X-original-commit: 3a9d5d829bb547632dd45c622f59684815c6eb62
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>