[FIX] sale_subscription: flush all to avoid test collisions
Issue:
------
Some tests of `TestSaleReport` introduced with the fix [^1], do not pass
if the module `website_sale` is installed.
This is due to the fact that the sale order flush does not flush all the
fields of the sale order line model that are used in the tests
(`price_subtotal` for example).

Solution:
---------
Use `flush_all` to flush the `sale.order` and `sale.order.line` models.

opw-3818584

[^1]: https://github.com/odoo/enterprise/commit/f886e65ce123f06ee8a5a2f149330864bcdb3254

closes odoo/enterprise#60042

X-original-commit: 3f64be1c207e327659eb6a422f92c073ab7ff654
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>