[FIX] *: add indexes
Deleting a partner may take a long time because odoo has to check an
entire table to find few records or no records at all that reference the
partner and set it to null.

So we are adding an index btree not null to speed up the deletion of
partners.

closes odoo/enterprise#60099

Task-id: 3759406
X-original-commit: 003834b42050dada590a431f71649f4adadc09ea
Related: odoo/odoo#160538
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>