[FIX] helpdesk_timesheet: fix typo in test causing indeterministic fails
Before this commit, the test changes the project_id inside a ticket with
the id of helpdesk team instead of altering the `team_id` field of that
ticket.

This commit fixes the issue to correctly change the `team_id` field
instead of `project_id` since the id given is the one of a helpdesk
team.

runbot-61010

closes odoo/enterprise#59955

X-original-commit: 65c1903c24f0d529ad32235ad97d9b1b2e6f594a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>