[FIX] account_followup: fix traceback while duplicating the followup levels
This traceback occurs when the user duplicate a followup level record.

To reproduce this issue:

1) Install `Accounting`
2) Open "Accounting/configuration/follow-up levels"
3) Select any one record and try to duplicate it.

Error:- "TypeError: argument of type 'NoneType' is not iterable"

When the user tries to duplicate any record, the default value is None,
which leads to the above traceback as there is no fallback for default.
Because of refactor in [1]

https://github.com/odoo/enterprise/blob/a1921fb9a27e5da3592dbc3d9c34e41776b6c8d4/account_followup/models/account_followup.py#L47-L51

After applying this commit, the issue will be resolved by giving a fallback value.

Additionally, an SQL unique name constraint was encountered when attempting to
duplicate it because of [2],so this commit also fix this issue by generating
new name by dading 15 days to the highest delay in that company + the record name
which is duplicating.

Previously duplicating several levels at the same time leads to days_uniq constraint
because of same delay.
This commit will allow to duplicate several levels at the same time by offsetting
highest existing delay + 15 arbitary days.

[1] https://github.com/odoo/enterprise/commit/e27a20d2e9a909260f28b18b0643b8205470236b
[2] https://github.com/odoo/enterprise/blob/278290dd220e7ca8d4b9909d1bc48a93ac14c0eb/account_followup/models/account_followup.py#L44

sentry-5134878308

closes odoo/enterprise#61404

X-original-commit: edf3b65651c4055b18c39ab2ea1b4efbb0c1a9c5
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>