[FIX] web_grid: fix sample data background in grid view
The background for the sample data has been removed in the following commit, thus it's not
visible in the grid, and the opacity of the record exceeds that of the view. Additionally, the
'o-sample-data-disabled' class is not set in the grid view and also is applied in another view
like Kanban. We have set the 'o-sample-data-disabled' class in the grid view  in this commit

Effect commit:
This commit resolves the issue with the background display of sample data in the grid view,
ensuring that the background is now visible and properly.

this commit is the refactoring of the content helper and sample data behavior-
https://github.com/odoo/odoo/commit/05819e8b4a4ebcb68f085353cd299303619cdbf0

task-3764514

closes odoo/enterprise#60023

X-original-commit: 74cb4591491f123fccee667d7f5b709c98f3dfd1
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>