[FIX] knowledge: support lower browser versions for checkURL
This commit fixes an issue with the CheckURL method in our utils, it wasn't
supporting low enough browser versions.

The static method canParse from the URL interface is only supported from
Firefox 115+, Chrome 120+ and Safari 17+.
This results in errors linked to the usage of this method, as it does not exist
in lower versions.

To fix that, this call is replaced by a try catch. If the URL constructor raises
a TypeError, this is expected and the function returns False. => This is the
same behavior as the static method version.

task-3842970

closes odoo/enterprise#60568

X-original-commit: 7c208e7733c4ff9ae9dba0d96eacf9449b7d9cf8
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>