[IMP] documents: add basic list view
This commit adds a basic list view for `documents.document`
model with no js_class attribute and with a higher priority
to make it usable by default in the activity view.
This avoid getting a fully customized list view with the
document inspector and drag and drop utilities which are
unnecessary.

Task ID : 3837270

Part-of: odoo/enterprise#61233