[FIX] sale_subscription: action_cancel reset the subscription_state
Before this commit, when a churned contract was cancelled, the
subscription_state was not reset. The resulting filter on SO would be
confusing:
Group by subscription_state > state

After this commit, the churned order subscription-state is reset to
draft or renewal quotation when canceled. This allow to reset the
canceled order to quotation and to be able to restart the flow if
necessary.

closes odoo/enterprise#59960

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>