[FIX] sale_subscription: action_cancel reset the subscription_state
Before this commit, when a churned contract was cancelled, the
subscription_state was not reset. The resulting filter on SO would be
confusing:
Group by subscription_state > state

After this commit, the churned order subscription-state is reset to
draft or renewal quotation when canceled. This allow to reset the
canceled order to quotation and to be able to restart the flow if
necessary.

closes odoo/enterprise#60134

X-original-commit: 3ec74fd2189dc431e3ea55a3c2142a422fffd0d5
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>