[REF] knowledge: refactor _insertEmbeddedView function
The function `_insertEmbeddedView` accepts 7 parameters to set up an
embedded view. Some of these parameters are optional or incompatible
with each other. For example: it makes no sense to pass the xml id of an
action and an actual action object. To insert an embedded view, the
developer has to pass many null, undefined or falsy values to the
function, making the code difficult to read and maintain.

To improve this, the function will accepts a single object containing
all the props of the embedded views. This object will directly be encoded
in the blueprint of the behavior. The caller will now be responsible for
providing a valid configuration with the required values.

The advantage of directly passing the props of the embedded view to the
`_insertEmbeddedView` function is that we will not have to update the
API of the function in the future whenever we define a new prop to the
embedded view behavior.

The `_insertEmbeddedView` function will also no longer copy the action
helper of the action to the behavior props of the embedded view. Doing
the copy was duplicating data and was not necessary as the action helper
is, by default, loaded from the loaded action unless a custom one is
defined in the behavior props.

task-3847699

Part-of: odoo/enterprise#51284