[FIX] account_online_synchronization: access denied to 'Read-only' users
In the system parameters, switch account_online_synchronization.proxy_mode to sandbox
Add a bank with plaid for example and follow the different steps
Have a user [DEMO] with Accounting access set to 'Read-only'
Log in with [DEMO]
Try to access the Accounting App Dashboard

Issue:
Access Error
"""
You are not allowed to access 'Bank Connection' (account.online.link) records.

This operation is allowed for the following groups:
        - Accounting/Accountant
        - Accounting/Bookkeeper

Contact your administrator to request access if necessary.
"""

This occurs because the read access to the account.online.link
and account.online.account records is granted only to users having at
least the account.group_account_user group
But, as those accesses are defined to be read only, it should be
greanted also to account.group_account_readonly users

opw-3821540

closes odoo/enterprise#60046

X-original-commit: 106f4c689d9c533027376a3172696c7db1816391
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>