[IMP] l10n_mx_edi_stock{_extended}: delivery guide PDF
The new version of the delivery guide includes changes in the PDF
format. New fields should be printed in the PDF file and other fields
should be removed.

We want to avoid having products in the delivery guide itself on the PDF
(only when it is really a carta-porte) because of security reasons
(driver does not necessarily need to know what's in their cargo).

In this commit, we create a new report "Carta Porte" that inherits from
the delivery slip report. Here, we remove all tables and added some
fields that are previously not present in the previous CCP version.

Task-id: 3755520
Part-of: odoo/enterprise#60006