[IMP] spreadsheet_edition: preserve filters through breadcrumb
Steps to reproduce:
- insert a pivot into a spreadsheet
- insert a global filter
- filter on a given value / or scroll / or select something
- right click on a pivot cell
- click on "See records"
- go back to the spreadsheet with the breadcrumb

=> the entire spreadsheet is reset. You lost the filter value,
the cells you selected, all data is loading again (which can take quite some
time)

Quoting our BA: "c'est un enfer".

Now the entire spreadsheet model is preserved.

Note that it's theoretically possible to "loop" over and have the same
spreadsheet open multiple times in the breadcrumb. It might be problematic
if the spreadsheet takes up a lot of memory. However, this scenario is unlikely
to happen in a real use case. At worst, the browser tab crashes out of memory,
but there's nothing too bad about it, just reload :)

A note on collaborative: when the user leaves the spreadsheet to "see records",
the collaborative channel is disconnected from the spreadsheet. But the messages
are still coming from other users and they are queued. When going back to the
spreadsheet, all queued messages are sent to the spreadsheet which can catch
up with the latest updates.

closes odoo/enterprise#60016

Task: 3810115
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>