[FIX] documents: don't apply `text-overflow` twice on the name
Issue
-----
The document name is replaced by `...` on the kanban card when the name
is too long and locking the document.

Steps to reproduce
------------------
- Install `documents` module
- Click on any document card to open the right sidebar
- Change the document name to a very long one (Should display `...` at
the end of name on the card)
- Click on the lock icon in the sidebar to lock the document

Cause
-----
There is a `o_text_overflow` CSS class on the field that use
`text-overflow` while a parent also use it(`o_kanban_record_title`).

Note: No issue before v16.0 because it seems that the renderer get ride
of any extra class on a text field in the kanban view.

Solution
--------
Remove the extra `o_text_overflow` CSS class from the field.

opw-3829011

closes odoo/enterprise#62133

X-original-commit: 1ab1e6b592edd9c1f48abd6c002207f74a96be2d
Signed-off-by: Nasreddin Boulif (bon) <bon@odoo.com>