[FIX] l10n_co_edi: hide EDI cancel button
Currently, when a Colombian invoice has been sent through the EDI, a
'Request EDI Cancellation' button appears on the form view. This button
shouldn't be there as the cancellations process is not allowed in
Colombia.

Steps to reproduce
------------------
* install `l10n_co_edi`
* switch to a Colombian company and set up the EDI
* create, confirm, and send an invoice
* once the invoice has been successfully sent, attempt to cancel it
  using the 'Request EDI Cancellation' button.

The invoice will be marked as `to cancel`. Odoo will then periodically
try to cancel it, but it will never succeed because Colombian invoices
that have been sent to the government cannot be reset to draft. If the
cancellation process is initiated by the user (either by manually
running the EDI services cron, or using the 'Process Now' button on the
form view) the following error message appears: "You can't edit the
following journal entry [journal name] because an electronic document
has already been sent to Carvajal. To edit this entry, you need to
create a Credit Note for the invoice and create a new invoice."

opw-3754395

closes odoo/enterprise#60015

X-original-commit: 99a1965583d04600b4fb50ce780fc757e04c10b0
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>