[IMP] web,*: add field registry validation
This is linked to its community commit: 69cdee2a222a

The bunch of lines modified in the test has a simple explaination:
To test the non existence of some options (eg: displayName), the code
would simply set the value to null after the spreading operation.
Like:
{
    ...charField, // contains a display name
    displayName: null,
}
With the registry validation, null is no longer a valid displayName
value. So to go around that in those couple tests, we remove the key
by deleting it.

Task id 3613319

closes odoo/enterprise#60080

Related: odoo/odoo#160343
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>