[IMP] mrp_workorder: add log notes on WO/MO
The user now can add log notes for WO/MO,
on shop floor and in the backend, via
MO miscellaneous tab.
MO Miscellaneous tab:
.go to Manufacturing order
.click on miscellaneous
.add a log note

MO View (shop floor):
.click on 3 dots
.click on Log Note
.a popup window is showing, add your
log note and save it.
Now, you can see it below the header of the MO.
If you click in that cell, you can edit the log note.

WO View (shop floor):
.click on 3 dots
.click on update instructions
.click on Log Note
.type your note, and save it.
Now, you can see it below the header of the
WO.
If you click in that cell, you can edit the log note.

The log note for a WO is the same for the MO.

Was added too a few improvement on the worksheet
sugestion button on mrp_menu_dialog. Now, it's
called Improvement Suggestion with the options
inside it:
.Update Instructions
.Delete a Step
.Add a Step

Task: 3648489
Part-of: odoo/enterprise#57534