[FIX] planning: fix traceback when the end date is false
This traceback occurs when the user makes the end date as empty
and tries to modify the start date.

To reproduce this issue:
1) Install planning
2) Create a new planning slot
3) click on any template id
4) Now remove the end date
5) A traceback occurs

Error:-
"TypeError: unsupported operand type(s) for -: 'bool' and 'datetime.datetime'"

When the user removes the 'end datetime' its value is False which leads to the
above traceback when - is used between start datime and end datetime.

https://github.com/odoo/enterprise/blob/5600b5b754f79e5d7a253fe9db45ac8a9273d3de/planning/models/planning.py#L406-L408

Additionally, a value error is also occurring when the start date-time is empty
while calculating start-end-dates. as it returns both datetime and timezone.

It only removes 'tzinfo' when the start date time and template is present.
https://github.com/odoo/enterprise/blob/5600b5b754f79e5d7a253fe9db45ac8a9273d3de/planning/models/planning.py#L621-L627

After applying this commit will resolve both issues and make code more robust

sentry-5005245970

closes odoo/enterprise#62318

X-original-commit: 63c12af2493a00c306d59e9c144f0e7549ef4861
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>